Added portfolio navigation
