+++ src/features/types/menu.ts
@@ -2,6 +2,7 @@ export enum MenuItems {
   HOME = "#hero",
   ABOUT = "#about",
   RESUME = "#resume",
+  PORTFOLIO = "#portfolio",
 }
 
 export type MenuType = MenuItems;
